updates testnet website info on uptime points
users need to be running the most recent version of Iron Fish to earn points for
hosting a node. there is a one week grace period to update. the website doesn't
say anything about this.

adds information on the one week grace period to the phase 3 'About' page, and
adds information to the FAQ answer around uptime points.